requirements.txt
azure-functions==1.21.3
openai==1.82.0
requests==2.32.3
nest_asyncio>=1.5.0
python-dotenv>=1.0.0
gremlinpython>=3.7.0
